requirements.txt
mcp
httpx
mangum
python-jose[cryptography]
